[#21] Refactor: move router to App.tsx
