+++ client/src/pages/index.ts
@@ -1,4 +1,7 @@
-export * from './Home';
+export * from './Root';
 export * from './SignUp';
 export * from './Login';
 export * from './Error';
+export * from './Edit';
+export * from './Todo';
+export * from './Default';
